Add files via upload
Today I did a lot better at writing the define function, and this one is for a Full Name Formatter. Essentially, this was for learning the real difference between what return does and what print does. "Return" really just returns the code back into the program for it to use later. "Print" will just show the result of the code. Very easy to understand after taking the practical approach to it. Below I added the defined function to a variable called "name" where I can print one name at a time.