Add files via upload
This task was the make_pizza() function, where I worked with *args. This was more interesting because instead of being stuck with a fixed number of parameters, the function can now take any number of toppings. Inside the function, a loop goes through the toppings and prints them out as part of the pizza summary. I really like this because it shows how flexible functions can be. You don’t need to hardcode things; you can pass in as much info as you want, and the function adjusts.